requirements.txt
accelerate==0.30.1
diffusers==0.29.2
torch==2.5.1
torchvision>=0.19.0
matplotlib
tensorboard==2.17.0
scikit-image==0.23.2
timm
transformers==4.30.2
deepspeed==0.14.4
open_clip_torch==2.22.0
pandas==2.2.2
opencv-python
gpustat
tqdm
ipdb
kornia
einops
trimesh
omegaconf
lpips
easydict
ninja
scipy
av
scikit-learn
pytorch-lightning==1.9.4
roma
imagesize
moviepy==1.0.3
plyfile
carvekit
pillow_heif